Use model classes in api module
